Add comment base abstract class
